restructure channel components
