+++ src/modules/channel/subcomponents/index.js
@@ -1,0 +1,1 @@
+export * from './navbar';
